requirements.txt
discord.py[voice]==1.3.3
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
pyyaml==5.3.1